+++ SoundflowerBed/Biquad.h
@@ -4,11 +4,11 @@
 
 class Biquad {
     protected:
-    float mX1, mX2;
-    float mY1, mY2;
-    float mB0, mB1, mB2, mA1, mA2;
-    float mB0dif, mB1dif, mB2dif, mA1dif, mA2dif;
-    float mInterpolationSteps;
+    double mX1, mX2;
+    double mY1, mY2;
+    double mB0, mB1, mB2, mA1, mA2;
+    double mB0dif, mB1dif, mB2dif, mA1dif, mA2dif;
+    int32_t mInterpolationSteps;
 
     void setCoefficients(int32_t steps, double a0, double a1, double a2, double b0, double b1, double b2);
